requirements.txt
tkintermapview
pillow
